#30 refactor: User 도메인 import 경로 정리
